pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name="witch"
version="7.0.0"
authors= [
{name ="Jack Orlowski-Scherer"},
{name ="Saianeesh Keshav Haridas"},
{name ="Luca Di Mascolo"},
]
description = "Fast modeling of galaxy clusters using JAX"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["galaxy clusters", "tSZ", "Sunyaev–Zeldovich", "MUSTANG2", "jax", "minkasi"]
license = {text = "GPLv3"}
classifiers = [ "Programming Language :: Python" ]
dependencies = ["numpy", "astropy", "jax", "jaxlib", "dill", "jitkasi @ git+https://github.com/MUSTANG-SZ/jitkasi.git", "cython", "mpi4py", "mpi4jax"]

[project.optional-dependencies]
fitter = ["pyyaml", "minkasi"]
profile = ["tensorflow", "gitpython"]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-homepage-copier",
]

[project.scripts]
witcher = "witch.fitter:main"

[tool.setuptools]
packages=["witch"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
